Revert "Move auth and monitoring to `__init__.py` of corresponding folders"
This reverts commit 98c7808a2b9f99469f39b7175296fd471f26c77a.
