+++ fractal_server/app/routes/monitoring.py
@@ -9,19 +9,19 @@ from fastapi import Depends
 from sqlalchemy import func
 from sqlmodel import select
 
-from ...db import AsyncSession
-from ...db import get_db
-from ...models import ApplyWorkflow
-from ...models import Dataset
-from ...models import JobStatusType
-from ...models import Project
-from ...models import Workflow
-from ...schemas import ApplyWorkflowRead
-from ...schemas import DatasetRead
-from ...schemas import ProjectRead
-from ...schemas import WorkflowRead
-from ...security import current_active_superuser
-from ...security import User
+from ..db import AsyncSession
+from ..db import get_db
+from ..models import ApplyWorkflow
+from ..models import Dataset
+from ..models import JobStatusType
+from ..models import Project
+from ..models import Workflow
+from ..schemas import ApplyWorkflowRead
+from ..schemas import DatasetRead
+from ..schemas import ProjectRead
+from ..schemas import WorkflowRead
+from ..security import current_active_superuser
+from ..security import User
 
 
 router_monitoring = APIRouter()
